[test] test with new version
