+++ test3.sh
@@ -1,0 +1,4 @@
+cc ./utils/vector.c ./utils/vector2.c ./utils/vector3.c ./utils/ray.c ./utils/color.c ./utils/geometry.c \
+	./yuikim/test3.c ./setting/view.c ./setting/check_hit.c ./utils/equation.c ./utils/min_func.c ./setting/cylinder.c \
+	./parsing/format_check.c ./parsing/ft_split.c ./parsing/ft_math.c ./parsing/parsing.c ./parsing/utility.c ./parsing/extract.c ./parsing/extract2.c ./parsing/ft_split_group.c\
+	./setting/check_world_hit.c ./utils/random.c ./utils/vector_random.c ./tracing/*.c ./utils/max_func.c
